Fixes to CPL_gdalrasterize and CPL_gdalnearblack
The fix to CPL_gdalnearblack is +/- self-explanatory: the driver kind
flag needs to communicate that we're needing raster rather than a
vector driver.

The fix to CPL_gdalrasterize reverts a breaking change made in commit
5d23803598f21043ff6ab7d64019cdd66f0cb4e0. That change led to failure
of calls to CPL_gdalrasterize with a message to the effect that "GDAL
Error 1: Attempt to create 0x0 dataset is illegal,sizes must be larger
than zero." My proposed edit repairs that and gets the function
working again.