fix read_stars for driver not well recognized
I had a case for which the raster file in ENVI format was not well recognised by sf::gdal_read. However, the data was well read with sf::gdal_read inputting a predefined driver='ENVI'. However, as `proxy` is now defined throw `is_big` in the signature, but without forwarding the driver. And `is_big` is calling read_stars thus leading to a wrong detection of the driver in my case. Forwarding the driver to is_big in the proxy definition of read_stars signature is avoiding such a behavior.